client in progress

+++ src/espnow_client.cpp
@@ -16,7 +16,7 @@ void setup() {
     
     config_reset_button();
     delay_ms(100);
-    if (is_reset_button_pressed()) {
+    if (false && is_reset_button_pressed()) {
         // start captive portal
         Logger::info("Force config portal");
         WiFiCaptivePortal::start_captive_portal(); // this is blocking until reboot
